moved functions from process_data.py to utilities.py to remove redundant module. Removed unused imports
