+++ recognition/ISIC-ImprovedUnet-44961374/utilities.py
@@ -2,9 +2,64 @@
 This script contains functions required for dsiplaying images and calculating DSC.
 @author: Mujibul Islam Dipto
 """
-import tensorflow as tf
-import matplotlib.pyplot as plt
+import tensorflow as tf # for DL functionalities
+import matplotlib.pyplot as plt # for plotting images
+from tensorflow.image import resize, decode_jpeg, decode_png # for handling jpg and png images
+from tensorflow.io import read_file # for reading image files
 
+
+def process_image(image, x, y):
+    """Proccesses the image of a scan. The image is decoded to a uint8 tensor.
+
+    Args:
+        image (Tensor): Tensor representing the image
+        x (float): x dimension for resizing
+        y (float): y dimension for resizing
+
+    Returns:
+        uint8 tensor: decoded image
+    """
+    image = decode_jpeg(image, channels=1)
+    image = resize(image, (x, y))
+    image = tf.cast(image, tf.float32) / 255.0
+    return image
+
+
+def process_mask(mask, x, y):
+    """Processes the image of a mask. The image is decoded to a uint8 or uint16 tensor.
+
+    Args:
+        mask (Tensor): Tensor represting the mask
+        x (float): x dimension for resizing
+        y (float): y dimension for resizing
+
+    Returns:
+        uint8 or uint16 tensor: decoded mask
+    """
+    mask = decode_png(mask, channels=1)
+    mask = resize(mask, (x, y))
+    return mask
+
+
+def process_data(image_path, mask_path):
+    """Processes an image and mask tuple. 
+
+    Args:
+        image_path (str): location of the image
+        mask_path (str): location of the mask
+
+    Returns:
+        tuple: decoded image and mask
+    """
+    # handle images of scans
+    image = read_file(image_path)
+    image = process_image(image, 256, 256)
+
+    # handle images of masks
+    mask = read_file(mask_path)
+    mask = process_mask(mask, 256, 256)
+    valid_mask = mask == [0, 255]
+    return image, valid_mask
 
 def display(data, figsize, cmap):
     """Displays a scan, mask and predicted mask of a lesion from given data.
@@ -25,8 +80,10 @@ def display(data, figsize, cmap):
 def dice_similarity(x, y):
     """Calculates the Dice Similarity Coefficient between two images, where:
     DSC = (2 * |x intersection y|) / |x| + |y|
-    More information about DSC can be found here: https://en.wikipedia.org/wiki/S%C3%B8rensen%E2%80%93Dice_coefficient
-    Function adapted from: https://medium.com/@karan_jakhar/100-days-of-code-day-7-84e4918cb72c
+    More information about DSC can be found here:
+    https://en.wikipedia.org/wiki/S%C3%B8rensen%E2%80%93Dice_coefficient
+    Function adapted from:
+    https://medium.com/@karan_jakhar/100-days-of-code-day-7-84e4918cb72c
     Args:
         x (Tensor): image 1
         y (Tensor): image 2
